Task14225

See merge request root/lyna!34

+++ web-app/src/main/java/com/lyna/web/domain/product/repository/impl/ProductRepositoryImpl.java
@@ -10,66 +10,59 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 import java.util.List;
 
 @Repository
-@Transactional
-public class ProductRepositoryImpl extends BaseRepository<Product, Long> implements ProductRepository {
+public class ProductRepositoryImpl extends BaseRepository<Product, String> implements ProductRepository {
 
     private final Logger log = LoggerFactory.getLogger(StoreRepositoryImpl.class);
 
-    @PersistenceContext
-    private EntityManager em;
 
     public ProductRepositoryImpl(EntityManager em) {
         super(Product.class, em);
     }
 
-    @Override
-    public void updateProduct(Product product) {
-        try {
-            String hql = "UPDATE Product p set p.tenantId = :tenantId, p.updateUser = :updateUser, p.updateDate = :updateDate,"
-                    + "p.code = :code,p.name = :name, p.unit = :unit, p.price = :price, p.category1 = :category1,"
-                    + "p.category2 = :category2, p.category3 = :category3 WHERE p.productId=:productId";
-            entityManager.createQuery(hql)
-                    .setParameter("tenantId", product.getTenantId())
-                    .setParameter("updateUser", product.getUpdateUser())
-                    .setParameter("updateDate", product.getUpdateDate())
-                    .setParameter("name", product.getName())
-                    .setParameter("unit", product.getUnit())
-                    .setParameter("price", product.getPrice())
-                    .setParameter("category1", product.getCategory1())
-                    .setParameter("category2", product.getCategory2())
-                    .setParameter("category3", product.getCategory3())
-                    .setParameter("productId", product.getProductId())
-                    .executeUpdate();
-        } catch (IllegalArgumentException e) {
-            log.error(e.getMessage());
-        }
-    }
 
     @Override
-    public Product findOneByProductId(String productId) {
+    @Transactional
+    public Product findOneByProductIdAndTenantId(String productId, int tenantId) {
         return entityManager
-                .createQuery("SELECT p FROM Product p WHERE p.productId=:productId", Product.class)
+                .createQuery("SELECT p FROM Product p WHERE p.productId=:productId AND P.tenantId=:tenantId", Product.class)
                 .setParameter("productId", productId)
+                .setParameter("tenantId", tenantId)
                 .getSingleResult();
     }
 
     @Override
-    public boolean deletebyProductId(List<String> listProductId) {
-        try {
-            String query = "DELETE FROM Product p WHERE p.productId in (:listProductId)";
-            entityManager.createQuery(query).setParameter("listProductId", listProductId).executeUpdate();
-            return true;
-        } catch (IllegalArgumentException e) {
-            log.error(e.getMessage());
-            throw e;
-        }
+    public boolean deleteByProductIdsAndTenantId(List<String> productIds, int tenantId) {
+
+        String query = "DELETE FROM Product p WHERE p.productId in (:listProductId) AND p.tenantId=:tenantId";
+        entityManager.createQuery(query)
+                .setParameter("listProductId", productIds)
+                .setParameter("tenantId", tenantId).executeUpdate();
+        return true;
     }
 
     @Override
+    @Transactional
+    public Product findOneByCodeAndTenantId(String code, int tenantId) {
+        return entityManager
+                .createQuery("SELECT p FROM Product p WHERE p.code=:code AND p.tenantId=:tenantId", Product.class)
+                .setParameter("code", code)
+                .setParameter("tenantId", tenantId)
+                .getSingleResult();
+
+    }
+
+    @Override
+    @Transactional
+    public List<Product> findByTenantId(int tenantId) {
+        return entityManager
+                .createQuery("SELECT p FROM Product p WHERE p.tenantId=:tenantId order by p.name", Product.class)
+                .setParameter("tenantId", tenantId).getResultList();
+    }
+
+    @Transactional
     public List<String> getListProductCodeByProductCode(int tenantId, List<String> products) {
         return entityManager
                 .createQuery("SELECT p.code FROM Product p WHERE p.tenantId = :tenantId and  p.code in (:products)")
@@ -79,6 +72,7 @@ public class ProductRepositoryImpl extends BaseRepository<Product, Long> impleme
     }
 
     @Override
+    @Transactional
     public List<Product> getProductsByProductCode(int tenantId, List<String> products) {
         return entityManager
                 .createQuery("SELECT p FROM Product p WHERE p.tenantId = :tenantId and  p.code in (:products)")
